Update file pathing to be cleaner
