--- src/components/Header.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export const Header = () => {
-  return (
-    <header className='main-header'>
-      <img className='logo' src='https://i.imgur.com/i6f9tQA.png' alt='onlyDevs logo' id='onlyDevsLogo' />
-    </header>
-  )
-}
